fix(commands): updates process empty command approach
